Update PHCpack.m2
When symbols in the original ring become part of a new, larger ring used inside a method,
then the symbols remain part of the larger ring once outside the method.
This is a nuisance for users of the method who do not want to make a new ring each time
they call the method.  Therefore, make a new ring as the original ring before leaving the method.